move llms plugin to a separate file within plugins directory
